chore: organize package structure
