+++ app/src/main/java/com/team22/soundary/MainActivity.kt
@@ -1,10 +1,10 @@
-package com.team22.soundary.feature.main
+package com.team22.soundary
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
-import com.team22.soundary.R
 import com.team22.soundary.databinding.ActivityMainBinding
+import com.team22.soundary.feature.main.presentation.MainFragment
 import com.team22.soundary.feature.profile.fragment.ProfileFragment
 import com.team22.soundary.feature.search.FriendSearchFragment
 import com.team22.soundary.feature.share.ShareMusicFragment
