+++ app/src/main/res/layout/fragment_main.xml
@@ -6,7 +6,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature.main.MainActivity">
+    tools:context=".MainActivity">
 
     <ImageView
         android:id="@+id/background_upper_imageView"
